# common/api/swagger.json
{
  "openapi": "3.0.0",
  "components": {
...
    "schemas": {
      "FieldErrors": {
        "properties": {},
        "type": "object",
        "additionalProperties": {
          "properties": {
            "value": {},
            "message": {
              "type": "string"
            }
          },
          "required": ["message"],
          "type": "object"
        }
      },
      "Pick_ValidateError.status-or-message-or-fields_": {
        "properties": {
          "status": {
            "type": "number",
            "format": "double"
          },
          "message": {
            "type": "string"
          },
          "fields": {
            "$ref": "#/components/schemas/FieldErrors"
          }
        },
        "required": ["status", "message", "fields"],
        "type": "object",
        "description": "From T, pick a set of properties whose keys are in the union K"
      },
      "Pick_Error.message_": {
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": ["message"],
        "type": "object",
        "description": "From T, pick a set of properties whose keys are in the union K"
      }
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "name": "access_token",
        "in": "query"
      }
    }
  },
  "info": {
    "title": "api",
    "version": "0.1.0",
    "contact": {}
  },
  "paths": {
...
    "/profile/{username}": {
      "get": {
        "operationId": "GetProfile",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "seekings": {
                      "items": {
                        "properties": {
                          "offer": {
                            "properties": {
                              "title": {
                                "type": "string"
                              }
                            },
                            "required": ["title"],
                            "type": "object"
                          }
                        },
                        "required": ["offer"],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "offers": {
                      "items": {
                        "properties": {
                          "offer": {
                            "properties": {
                              "title": {
                                "type": "string"
                              }
                            },
                            "required": ["title"],
                            "type": "object"
                          }
                        },
                        "required": ["offer"],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "lastName": {
                      "type": "string"
                    },
                    "firstName": {
                      "type": "string"
                    },
                    "academicTitle": {
                      "type": "string"
                    },
                    "bio": {
                      "type": "string"
                    },
                    "areas": {
                      "items": {
                        "properties": {
                          "area": {
                            "properties": {
                              "name": {
                                "type": "string"
                              }
                            },
                            "required": ["name"],
                            "type": "object"
                          }
                        },
                        "required": ["area"],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "email": {
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    },
                    "background": {
                      "type": "string"
                    },
                    "avatar": {
                      "type": "string"
                    },
                    "url": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "seekings",
                    "offers",
                    "lastName",
                    "firstName",
                    "academicTitle",
                    "bio",
                    "areas",
                    "email",
                    "id",
                    "background",
                    "avatar",
                    "url"
                  ],
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pick_ValidateError.status-or-message-or-fields_"
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "status": 401,
                      "message": "Authentication failed",
                      "fields": {
                        "access_token": {
                          "message": "Invalid access token"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Pick_Error.message_"
                    },
                    {
                      "properties": {
                        "status": {
                          "type": "number",
                          "format": "double"
                        }
                      },
                      "required": ["status"],
                      "type": "object"
                    }
                  ]
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "status": 500,
                      "message": "Internal Server Error"
                    }
                  }
                }
              }
            }
          }
        },
        "description": "Retrieve a profile by username of the community including their public information.",
        "summary": "Retrieve profile by username.",
        "tags": ["Profile"],
        "security": [
          {
            "api_key": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/event/{slug}": {
      "get": {
        "operationId": "GetEvent",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "tags": {
                      "items": {
                        "properties": {
                          "tag": {
                            "properties": {
                              "title": {
                                "type": "string"
                              }
                            },
                            "required": ["title"],
                            "type": "object"
                          }
                        },
                        "required": ["tag"],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "stage": {
                      "properties": {
                        "title": {
                          "type": "string"
                        }
                      },
                      "required": ["title"],
                      "type": "object"
                    },
                    "experienceLevel": {
                      "properties": {
                        "title": {
                          "type": "string"
                        }
                      },
                      "required": ["title"],
                      "type": "object"
                    },
                    "participationFrom": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "canceled": {
                      "type": "boolean"
                    },
                    "venueZipCode": {
                      "type": "string"
                    },
                    "venueCity": {
                      "type": "string"
                    },
                    "venueStreetNumber": {
                      "type": "string"
                    },
                    "venueStreet": {
                      "type": "string"
                    },
                    "venueName": {
                      "type": "string"
                    },
                    "participationUntil": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "parentEventId": {
                      "type": "string"
                    },
                    "endTime": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "startTime": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "types": {
                      "items": {
                        "properties": {
                          "eventType": {
                            "properties": {
                              "title": {
                                "type": "string"
                              }
                            },
                            "required": ["title"],
                            "type": "object"
                          }
                        },
                        "required": ["eventType"],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "focuses": {
                      "items": {
                        "properties": {
                          "focus": {
                            "properties": {
                              "title": {
                                "type": "string"
                              }
                            },
                            "required": ["title"],
                            "type": "object"
                          }
                        },
                        "required": ["focus"],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "areas": {
                      "items": {
                        "properties": {
                          "area": {
                            "properties": {
                              "name": {
                                "type": "string"
                              }
                            },
                            "required": ["name"],
                            "type": "object"
                          }
                        },
                        "required": ["area"],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "targetGroups": {
                      "items": {
                        "properties": {
                          "targetGroup": {
                            "properties": {
                              "title": {
                                "type": "string"
                              }
                            },
                            "required": ["title"],
                            "type": "object"
                          }
                        },
                        "required": ["targetGroup"],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "participantLimit": {
                      "type": "number",
                      "format": "double"
                    },
                    "description": {
                      "type": "string"
                    },
                    "subline": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    },
                    "background": {
                      "type": "string"
                    },
                    "url": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "tags",
                    "stage",
                    "experienceLevel",
                    "participationFrom",
                    "canceled",
                    "venueZipCode",
                    "venueCity",
                    "venueStreetNumber",
                    "venueStreet",
                    "venueName",
                    "participationUntil",
                    "parentEventId",
                    "endTime",
                    "startTime",
                    "types",
                    "focuses",
                    "areas",
                    "targetGroups",
                    "participantLimit",
                    "description",
                    "subline",
                    "name",
                    "id",
                    "background",
                    "url"
                  ],
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pick_ValidateError.status-or-message-or-fields_"
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "status": 401,
                      "message": "Authentication failed",
                      "fields": {
                        "access_token": {
                          "message": "Invalid access token"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Pick_Error.message_"
                    },
                    {
                      "properties": {
                        "status": {
                          "type": "number",
                          "format": "double"
                        }
                      },
                      "required": ["status"],
                      "type": "object"
                    }
                  ]
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "status": 500,
                      "message": "Internal Server Error"
                    }
                  }
                }
              }
            }
          }
        },
        "description": "Retrieve a event by slug of the community including their public information.",
        "summary": "Retrieve event by slug.",
        "tags": ["Event"],
        "security": [
          {
            "api_key": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}